Rework LR's polling model to accurately reflect orders/listings

+++ scripts/start.js
@@ -14,12 +14,9 @@ import {
   createNFTClient,
 } from "../src/blockchain/index.js";
 import logError from "../src/log-error.js";
-import sleep from "../src/sleep.js";
 import { createDbClient } from "../src/database/index.js";
 import { getDefaultProvider } from "ethers";
 import createBotClient from "../src/discord/create-bot-client.js";
-// eslint-disable-next-line no-unused-vars
-import { Client } from "discord.js";
 
 dotenv.config({ path: path.resolve(".env") });
 
@@ -31,8 +28,13 @@ const {
   ALCHEMY_API_KEY,
 } = process.env;
 
-const EVENT_LISTENER_REFRESH_PERIOD = 5 * 60 * 1000;
 const POLL_USER_TOKENS_INTERVAL = 5 * 60 * 1000;
+
+/*
+ * Every "N" NFT polls, the Discord client is reset to make sure
+ * it doesn't die silently.
+ */
+const POLLS_PER_RESET = 10;
 
 const ethProvider = getDefaultProvider("homestead", {
   etherscan: ETHERSCAN_API_KEY,
@@ -51,8 +53,6 @@ const ethProvider = getDefaultProvider("homestead", {
  * @param {Object} params
  * @param {Object} params.dbClient - The initialized database client.
  * @param {Object} params.nftClient - The initialized client to poll NFT
- * @param {EventEmitter} params.nftEventEmitter - The initialized client to
- * poll NFT on-chain events.
  */
 const monitorBlockchainEvents = async ({
   dbClient,
@@ -86,14 +86,13 @@ const monitorBlockchainEvents = async ({
 
   /*
    * Creates a map where each collection address has associated a list of
-   * watchers (users and servers), token ids, current offer price and expiry.
+   * watchers (users and servers) and their token ids.
    * The purpose of this map is to pull information only for the collections
    * which are relevant to users.
    */
   const createCollectionMap = async () => {
     const { objects: alerts } = await dbClient.getAllAlerts();
     const updatedAlerts = await updateAlertTokens(alerts);
-    const { objects: offers } = await dbClient.getAllCollectionOffers();
     const collectionMap = updatedAlerts.reduce(
       (collectionMap, { id, tokens, ...alert }) => {
         const userCollections = tokens.reduce((collections, token) => {
@@ -110,66 +109,44 @@ const monitorBlockchainEvents = async ({
             watchers: currentWatchers.concat([{ ...alert, id, tokenIds }]),
           };
         });
-        // Set collection offers
-        offers.forEach(({ collection, price, endsAt }) => {
-          const current = collectionMap[collection];
-          if (current == null) {
-            return;
-          }
-
-          current.highestOffer = price;
-          current.endsAt = endsAt;
-          collectionMap[collection] = current;
-        });
         return collectionMap;
       },
       {}
     );
-    const floors = await Promise.all(
-      Object.keys(collectionMap).map((collection) =>
-        dbClient.getCollectionFloor(collection).then(({ object }) => object)
-      )
-    );
-    // Set collection floors
-    floors
-      .filter((floor) => floor != null)
-      .forEach(({ collection, price }) => {
-        const current = collectionMap[collection];
-        if (current == null) {
-          return;
-        }
-
-        current.collectionFloor = price;
-        collectionMap[collection] = current;
-      });
 
     return collectionMap;
   };
 
   const collectionMap = await createCollectionMap();
-  nftEventEmitter.setCollections(Object.keys(collectionMap));
-  nftEventEmitter.removeAllListeners();
+  const collections = Object.keys(collectionMap);
   const botClient = await createBotClient({
     dbClient,
     nftClient,
   });
+  nftEventEmitter.poll(collections);
+  nftEventEmitter.start();
   nftEventEmitter.on("event", (event) => {
-    const entry = collectionMap[event.collection];
-    if (entry) {
-      event.highestOffer = entry.highestOffer || 0;
-      event.highestOfferEndsAt = entry.offerEndsAt;
-      event.collectionFloor = entry.collectionFloor;
-      event.watchers = entry.watchers;
-    }
-
-    botClient.emit("nftEvent", event);
+    const entry =
+      event.collection == null
+        ? {}
+        : collectionMap[event.collection.toLowerCase()];
+    const { watchers = [] } = entry || {};
+    botClient.emit("nftEvent", { ...event, watchers });
   });
-  await sleep(EVENT_LISTENER_REFRESH_PERIOD);
-  botClient.destroy();
-  return monitorBlockchainEvents({
-    dbClient,
-    nftClient,
-    nftEventEmitter,
+  let currentPolls = 0;
+  nftEventEmitter.on("pollEnded", () => {
+    if (currentPolls < POLLS_PER_RESET) {
+      currentPolls += 1;
+      nftEventEmitter.poll();
+    } else {
+      nftEventEmitter.stop();
+      botClient.destroy();
+      monitorBlockchainEvents({
+        dbClient,
+        nftClient,
+        nftEventEmitter,
+      });
+    }
   });
 };
 
@@ -183,6 +160,7 @@ const start = async () => {
 start();
 
 process.on("unhandledRejection", (error) => {
+  console.log(error);
   logError(`Unhandled promise rejection: ${error.toString()}`);
   process.exit(-1);
 });
